refactoring in prep for dockerfile changes to support workers
